Safari-609.1.11.1/LayoutTests/accessibility/mac/live-region-on-span-expected.txt

text
This tests that if aria-live is on a , that that is not ignored, and that is has the appropriate aria live information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS divElement.childAtIndex(0).role is 'AXRole: AXGroup'
PASS divElement.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXStaticText'
PASS divElement.childAtIndex(0).isAttributeSupported('AXARIALive') is true
PASS divElement.childAtIndex(0).isAttributeSupported('AXARIARelevant') is true
PASS successfullyParsed is true

TEST COMPLETE
